INGK-1290 Move TSN code to its own folder
