+++ ingenialink/ethernet/tsn/__init__.py
@@ -1,0 +1,1 @@
+"""Time-Sensitive Networking protocol support."""
